backend/requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-cors-headers==3.1.1
django-filter==2.0.0
django-rest-registration==0.3.12
djangorestframework==3.8.2
drf-yasg==1.10.2
flex==6.13.2
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
Markdown==3.0.1
MarkupSafe==1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
PyYAML==3.13
requests==2.19.1
rfc3987==1.3.8
ruamel.yaml==0.15.72
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.1
uritemplate==3.0.0
urllib3==1.23
validate-email==1.3
